Authentication works with the task owner being set autimatically, next up is user avatar & task images.

+++ backend/projectplanner_backend/src/task/task.service.ts
@@ -6,12 +6,17 @@ import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm/dist';
 import { HttpException } from '@nestjs/common/exceptions';
 import { HttpStatus } from '@nestjs/common/enums';
+import { TaskModule } from './task.module';
+import { User } from 'src/user_auth/entity/user';
 
 @Injectable()
 export class TaskService {
 
 
-    constructor(@InjectRepository(Task) private taskRepository: Repository<Task>) {}
+    constructor(
+        @InjectRepository(Task) private taskRepository: Repository<Task>,
+        @InjectRepository(User) private userRepository: Repository<User>
+    ) {}
     
     async getTasks(): Promise<Task[]> {
         return this.taskRepository.find();
@@ -25,15 +30,41 @@ export class TaskService {
         return task;
     }
 
-    async createTask(task: TaskDto): Promise<void>{
-        await this.taskRepository.save(task);
+    async createTask(taskDto: TaskDto, userName: string): Promise<void>{
+        const owner = await this.userRepository.findOneOrFail({
+            where: {username: userName}
+        })
+
+        await this.taskRepository.save({...taskDto, owner});
     }
 
-    async updateDesription(id: string, description: string){
+    async updateDesription(id: string, description: string, userid: string){
+        const task = await this.taskRepository.findOneOrFail({
+            where: {id},
+            relations: ['owner']
+        });
+
+        if (task.owner.id !== userid){
+            throw new HttpException(
+                "Unauthorized", 400
+            )
+        }
         await this.taskRepository.update({id}, {description});
     }
 
-    async deleteTask(id: string){
+    async deleteTask(id: string, userName: string){
+
+        const task = await this.taskRepository.findOneOrFail({
+            where: {id},
+            relations: ['owner']
+        });
+
+        if (task.owner.username !== userName){
+            throw new HttpException(
+                "You're not the owner of this task", 400
+            )
+        }
+
         this.taskRepository.delete({id});
 
     }
